\name{f_DB_cols}
\alias{f_DB_cols}
\title{Tables names in the table from a DataBase}
\usage{
f_DB_cols(table, 'txt')
}
\arguments{
\item{table}{A table in the current DB.}
\item{txt}{Optional: A string.}
}
\value{
}
\description{
Returns the list of columns in the tables from the DataBase.\cr
If we specify a txt, it will return only the column names that contain the txt.
}
\examples{
}
