\name{f_table}
\alias{f_table}
\title{Pring an histogram from a column}
\usage{
f_table(column)
}
\arguments{
\item{column}{A column of a dataframe.}
}
\value{
}
\description{
Retruns a df with the unique values of a clumns and the number of occurence of each unique values.
}
\examples{
f_table(mpg$hwy)
}
